Project DataWarehouse/pyproject.toml
[project]
name = "project-datawarehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "bs4>=0.0.2",
    "ddgs>=9.12.0",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "pandas>=3.0.2",
    "plotly>=6.6.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
    "supabase>=2.28.3",
]
